+++ src/common/game/weapons/meleeweapon.h
@@ -25,6 +25,6 @@ public:
         const Stats& stats
     );
 
-    void update(uint32_t timeSinceLastFrame);
+    void update(int64_t timeSinceLastFrame);
     Type getType(void) const;
 };
